refactor(core): extract fork-height accessors into own module
blockchain.rs had grown to ~4k lines partly because the 12
consensus-fork height readers (~324 lines) were inlined mid-file
between unrelated concerns. They share one shape — read env var,
parse, fall back to u64::MAX — and have no dependency on the
Blockchain struct or any local state. Pure config readers.

Pull them into crate::fork_heights so the concern lives in one place
and blockchain.rs gets ~10% smaller. Everything else identical.

API stability: blockchain.rs re-exports every moved symbol via
pub use, so existing import paths still resolve unchanged:
- crate::blockchain::get_voyager_fork_height
- crate::blockchain::get_evm_fork_height
- ... (12 total + warn_if_jail_consensus_armed)

Verified no caller change needed. Only two external call sites
exist (block_executor.rs + sentrix-rpc/explorer/layout.rs) and
both keep working bit-identically through the re-export.

get_chain_id stays in blockchain.rs because it couples to CHAIN_ID
const, scoped for a later split.

Tests: kept the existing fork-height test (test_jail_consensus_*)
coverage via blockchain.rs's mod tests. Added two new fork_heights
unit tests in the new module — one round-trip for the env-read
pattern (VOYAGER_FORK_HEIGHT chosen as representative; the other
11 readers are byte-identical implementations of the same shape)
and one smoke test for warn_if_jail_consensus_armed across both
armed and disarmed paths.

cargo check + clippy --tests -D warnings clean.
sentrix-core test suite: 215 passed (was 213; +2 new fork_heights tests).
Criterion bench unchanged (refactor-only, no hot-path code touched).
