+++ crates/sentrix-core/src/blockchain.rs
@@ -47,330 +47,16 @@ pub const BLOCK_TIME_SECS: u64 = 1;
 pub const MAX_TX_PER_BLOCK: usize = 5000;
 pub const CHAIN_ID: u64 = 7119; // default; overridable via SENTRIX_CHAIN_ID env
 
-/// Default Voyager DPoS fork activation height.
-/// u64::MAX = disabled (Pioneer-only). Override via VOYAGER_FORK_HEIGHT env var.
-const VOYAGER_DPOS_HEIGHT_DEFAULT: u64 = u64::MAX;
-
-/// Default Voyager EVM activation height.
-/// u64::MAX = disabled. Override via VOYAGER_EVM_HEIGHT env var.
-const VOYAGER_EVM_HEIGHT_DEFAULT: u64 = u64::MAX;
-
-/// V4 Step 3 / reward-v2 hard-fork height — coinbase routes to
-/// `PROTOCOL_TREASURY` at/after this height, ClaimRewards dispatch
-/// becomes valid. u64::MAX = disabled (Pioneer + v2.1.x accumulator-only
-/// behaviour). Override via `VOYAGER_REWARD_V2_HEIGHT` env var.
-///
-/// This is a CONSENSUS CHANGE. Every validator on the same chain must
-/// set the same value; mismatch produces a fork. Coordinated operator
-/// rollout required.
-const VOYAGER_REWARD_V2_HEIGHT_DEFAULT: u64 = u64::MAX;
-
-/// Tokenomics v2 fork height — at/after this block, `MAX_SUPPLY` becomes
-/// `MAX_SUPPLY_V2` (315M) and `HALVING_INTERVAL` becomes `HALVING_INTERVAL_V2`
-/// (126M, BTC-parity 4-year cadence). Pre-fork blocks retain the v1 schedule
-/// (42M halving, 210M cap). u64::MAX = disabled. Override via
-/// `TOKENOMICS_V2_HEIGHT` env var.
-///
-/// This is a CONSENSUS CHANGE — emission curve diverges across the fork.
-/// Every validator must set the same value; mismatch produces a fork.
-/// Coordinated operator rollout required (testnet bake first).
-///
-/// Fork should be activated while still in v1 era 0 (height < 42M) so the
-/// halving-count transition is smooth (both schedules give 0 halvings at
-/// fork moment, no reward jump). Activating after era 0 boundary would
-/// require additional dispatch logic to preserve cumulative halvings.
-const TOKENOMICS_V2_HEIGHT_DEFAULT: u64 = u64::MAX;
-
-/// BFT-gate-relax fork height — at/after this block, the validator-loop's
-/// "P1 BFT safety gate" relaxes from `active >= MIN_BFT_VALIDATORS (=4)`
-/// to `active >= ⌈2/3 × total_validator_count⌉`. For our 4-validator
-/// network this drops the gate threshold from 4 to 3, allowing BFT to
-/// continue when 1 validator is locally-jailed (= jail-cascade liveness
-/// margin).
-///
-/// SAFETY: BFT supermajority for finality is `⌈2/3 × N⌉` votes. With
-/// active=3 of total=4 and all 3 active validators sign, the precommit
-/// threshold (3 of 4 stake-weighted) is reached → finality possible.
-/// With active=2 of total=4, 2 < 3 → gate still blocks (correct).
-///
-/// This is a CONSENSUS-LIVENESS CHANGE (not safety-affecting if peers
-/// converge on the new threshold); activate via env var fork pattern.
-/// Fork is OPTIONAL — leaves as `u64::MAX` until operator decides to
-/// flip on. See `audits/jail-cascade-root-cause-analysis.md`.
-const BFT_GATE_RELAX_HEIGHT_DEFAULT: u64 = u64::MAX;
-
-/// Phase B (consensus-computed jail) fork height. Activates the
-/// `StakingOp::JailEvidenceBundle` dispatch path: epoch-boundary
-/// proposer includes JailEvidence in block, peers Pass-1-validate by
-/// recomputing from chain history, jail decision applied as consensus
-/// state mutation (deterministic by design).
-///
-/// Pre-fork: legacy `SlashingEngine::check_liveness` runs at epoch
-/// boundary (per-validator, locally-computed jail). Post-fork: Phase B
-/// dispatch takes over (consensus-applied jail).
-///
-/// CONSENSUS CHANGE — every validator must set the same value;
-/// mismatch produces a fork. Coordinated operator rollout required.
-/// u64::MAX = disabled (safe default while Phase B implementation
-/// is incomplete). Wire-format stable per Phase A (PR #359).
-const JAIL_CONSENSUS_HEIGHT_DEFAULT: u64 = u64::MAX;
-
-/// Activation height for SRC-721 + SRC-1155 native NFT/multi-token TokenOp
-/// variants. Pre-fork: dispatch rejects (`is_nft_family()` returning true).
-/// Post-fork: full handlers run (storage layer + REST).
-///
-/// CONSENSUS CHANGE — every validator must set the same value; mismatch
-/// produces a fork. Operator activates with halt-all + simultaneous-start
-/// after testnet bake.
-/// u64::MAX = disabled (safe default; wire format stable from this PR).
-const NFT_TOKENOP_HEIGHT_DEFAULT: u64 = u64::MAX;
-
-/// Activation height for `StakingOp::AddSelfStake` dispatch. Lets a
-/// validator's wallet bond real SRX into its own `self_stake` without
-/// the phantom-mint that `force-unjail` produces. Designed as the
-/// proper recovery path for slashed validators whose `self_stake <
-/// MIN_SELF_STAKE` (the 2026-04-27 self-stake-shortfall incident).
-///
-/// Pre-fork: dispatch rejects (wire format stable, dispatch gated).
-/// Post-fork: tx.amount transferred validator → treasury via the
-/// outer apply-Pass-2 transfer; `self_stake` incremented in registry.
-/// Supply-invariant preserving — no mint.
-///
-/// CONSENSUS CHANGE — every validator must set the same value;
-/// mismatch produces a fork. Operator activates with halt-all +
-/// simultaneous-start after testnet bake.
-/// u64::MAX = disabled (safe default; wire format stable from this PR).
-const ADD_SELF_STAKE_HEIGHT_DEFAULT: u64 = u64::MAX;
-
-/// EVM value-transfer activation height. Pre-fork: Pass-2 EVM apply path
-/// runs every tx with `TxEnv.value = U256::ZERO` regardless of envelope
-/// value (matches v2.1.48 behaviour — value-bearing EVM txs no-op the
-/// transfer but native fee still debits via Pass-1 `charge_fee_only`).
-/// Post-fork: revm sees the real envelope value and moves SRX between
-/// EOAs / forwards into payable contract calls.
-///
-/// Why gated: shipped flat in v2.1.49, recurred the eager-write divergence
-/// pattern that v2.1.48's FinalizeBlock guard was meant to close. Three
-/// 2v2 split-brain halts on 2026-05-01 (h≈1180k / 1191k / 1192k) all
-/// followed the same shape — validator-pair A finalizes one hash,
-/// validator-pair B the other. Faction split was deterministic across
-/// recoveries, suggesting the value-transfer apply path produces
-/// validator-specific divergence (likely revm/wei-sentri rounding
-/// interaction or block-context non-determinism — RCA pending in
-/// `audits/2026-05-01-evm-value-transfer-divergence.md`).
-///
-/// Until that RCA lands, default disabled mirrors v2.1.48 behaviour.
-/// Operator activates with halt-all + simul-start after the bug is
-/// understood + a regression test pins the invariant.
-const EVM_VALUE_TRANSFER_HEIGHT_DEFAULT: u64 = u64::MAX;
-
-/// Read Voyager fork height from env, default u64::MAX (mainnet safe).
-/// Testnet sets VOYAGER_FORK_HEIGHT=<height> in systemd service.
-pub fn get_voyager_fork_height() -> u64 {
-    std::env::var("VOYAGER_FORK_HEIGHT")
-        .ok()
-        .and_then(|v| v.parse().ok())
-        .unwrap_or(VOYAGER_DPOS_HEIGHT_DEFAULT)
-}
-
-/// Read EVM fork height from env, default u64::MAX (disabled).
-/// Testnet: set VOYAGER_EVM_HEIGHT=<height> in systemd service.
-pub fn get_evm_fork_height() -> u64 {
-    std::env::var("VOYAGER_EVM_HEIGHT")
-        .ok()
-        .and_then(|v| v.parse().ok())
-        .unwrap_or(VOYAGER_EVM_HEIGHT_DEFAULT)
-}
-
-/// V4 Step 3: read reward-v2 hard-fork height from env, default
-/// u64::MAX (disabled — keeps current accumulator-only behaviour).
-/// Post-fork: coinbase → `PROTOCOL_TREASURY`, ClaimRewards dispatch
-/// becomes consensus-valid.
-pub fn get_reward_v2_fork_height() -> u64 {
-    std::env::var("VOYAGER_REWARD_V2_HEIGHT")
-        .ok()
-        .and_then(|v| v.parse().ok())
-        .unwrap_or(VOYAGER_REWARD_V2_HEIGHT_DEFAULT)
-}
-
-/// Tokenomics v2: read fork height from env, default u64::MAX (disabled —
-/// keeps v1 emission schedule: 42M halving + 210M cap). Post-fork:
-/// 126M halving (BTC-parity 4-year) + 315M cap.
-pub fn get_tokenomics_v2_height() -> u64 {
-    std::env::var("TOKENOMICS_V2_HEIGHT")
-        .ok()
-        .and_then(|v| v.parse().ok())
-        .unwrap_or(TOKENOMICS_V2_HEIGHT_DEFAULT)
-}
-
-/// BFT-gate-relax: read fork height from env, default `u64::MAX`
-/// (disabled — keeps current `active >= MIN_BFT_VALIDATORS` gate).
-/// Post-fork: `active >= ⌈2/3 × total⌉` (= 3 for 4-validator network).
-/// Fork is optional — operators set `BFT_GATE_RELAX_HEIGHT=<height>`
-/// when they want to enable jail-cascade liveness margin.
-/// See `audits/jail-cascade-root-cause-analysis.md`.
-pub fn get_bft_gate_relax_height() -> u64 {
-    std::env::var("BFT_GATE_RELAX_HEIGHT")
-        .ok()
-        .and_then(|v| v.parse().ok())
-        .unwrap_or(BFT_GATE_RELAX_HEIGHT_DEFAULT)
-}
-
-/// Phase B: read JAIL_CONSENSUS_HEIGHT from env. Default `u64::MAX`
-/// (disabled). Activates consensus-computed jail dispatch when set.
-pub fn get_jail_consensus_height() -> u64 {
-    std::env::var("JAIL_CONSENSUS_HEIGHT")
-        .ok()
-        .and_then(|v| v.parse().ok())
-        .unwrap_or(JAIL_CONSENSUS_HEIGHT_DEFAULT)
-}
-
-/// Startup-time guardrail: if `JAIL_CONSENSUS_HEIGHT` is set to anything
-/// other than `u64::MAX`, log a loud warning explaining what's at stake.
-///
-/// Why this exists: the consensus-jail dispatch has a known LivenessTracker
-/// non-determinism bug that has halted mainnet twice (h=892799 on 2026-04-29,
-/// h=979199 on 2026-04-30). Each incident the recovery loop ends with
-/// "set JAIL_CONSENSUS_HEIGHT=u64::MAX in env" — but env files are easy to
-/// forget about and easy to revert by accident. This warning fires loudly
-/// at every startup so the operator can catch a misconfig BEFORE the next
-/// epoch boundary triggers the divergence.
-///
-/// Removing this warning is tied to landing the actual canonical-only
-/// LivenessTracker fix (see `audits/AUDIT_REPORT_2026_04_29.md` Phase 3).
-/// Until then, any value other than u64::MAX is operator risk.
-pub fn warn_if_jail_consensus_armed() {
-    let height = get_jail_consensus_height();
-    if height < u64::MAX {
-        tracing::warn!(
-            "⚠️  JAIL_CONSENSUS_HEIGHT={} (NOT u64::MAX) — consensus-jail \
-             dispatch ARMED. Known to halt mainnet via LivenessTracker \
-             non-determinism (incidents: h=892799 / 2026-04-29, h=979199 / \
-             2026-04-30). Set env to 18446744073709551615 unless you have \
-             explicitly verified the canonical-only LivenessTracker fix \
-             has shipped. See internal Sentrix Labs audit (2026-04-29).",
-            height
-        );
-        // Eprintln backup so the warning is also visible without RUST_LOG
-        // wired to WARN — defensive against ops accidentally running with
-        // info-only logging.
-        eprintln!(
-            "⚠️  JAIL_CONSENSUS_HEIGHT={} — see warn log above. This is a \
-             KNOWN HALT CONDITION. Confirm intent before continuing.",
-            height
-        );
-    }
-}
-
-/// Read NFT TokenOp fork height from env, default `u64::MAX` (disabled).
-/// Post-fork: SRC-721 + SRC-1155 dispatch active. Operators activate via
-/// halt-all + simultaneous-start with `NFT_TOKENOP_HEIGHT=<height>`.
-pub fn get_nft_tokenop_height() -> u64 {
-    std::env::var("NFT_TOKENOP_HEIGHT")
-        .ok()
-        .and_then(|v| v.parse().ok())
-        .unwrap_or(NFT_TOKENOP_HEIGHT_DEFAULT)
-}
-
-/// Read AddSelfStake fork height from env, default `u64::MAX` (disabled).
-/// Post-fork: `StakingOp::AddSelfStake` dispatch active — validators can
-/// top up their own `self_stake` with real SRX. Operators activate via
-/// halt-all + simultaneous-start with `ADD_SELF_STAKE_HEIGHT=<height>`.
-pub fn get_add_self_stake_height() -> u64 {
-    std::env::var("ADD_SELF_STAKE_HEIGHT")
-        .ok()
-        .and_then(|v| v.parse().ok())
-        .unwrap_or(ADD_SELF_STAKE_HEIGHT_DEFAULT)
-}
-
-/// Read EVM value-transfer fork height from env, default `u64::MAX`
-/// (disabled — matches v2.1.48 EVM behaviour). Post-fork: Pass-2 EVM
-/// apply path threads envelope value into `TxEnv.value` so revm performs
-/// real SRX transfers between EOAs and into payable contract calls.
-pub fn get_evm_value_transfer_height() -> u64 {
-    std::env::var("EVM_VALUE_TRANSFER_HEIGHT")
-        .ok()
-        .and_then(|v| v.parse().ok())
-        .unwrap_or(EVM_VALUE_TRANSFER_HEIGHT_DEFAULT)
-}
-
-/// Audit H3 (2026-05-06): EVM gas-fix fork height. Pre-fork the
-/// write-path EVM tx flow let revm internally deduct
-/// `gas_used × INITIAL_BASE_FEE` from sender's wei balance, and the
-/// writeback's wei→sentri floor-div silently dropped that delta (not
-/// burned, not credited to validator) — `total_supply()` leaked.
-/// Post-fork we set `cfg.disable_base_fee = true` on the write path
-/// too so revm skips gas accounting; Pass-1 native `tx.fee` (10K
-/// sentri flat) is the entire fee, with the existing 50/50
-/// burn/validator split intact. Default `u64::MAX` keeps the leak
-/// behaviour unchanged on chains that haven't activated; operator
-/// flips to a real height after testnet bake confirms the
-/// supply-conservation invariant on real flows.
-const EVM_GAS_FIX_HEIGHT_DEFAULT: u64 = u64::MAX;
-pub fn get_evm_gas_fix_height() -> u64 {
-    std::env::var("EVM_GAS_FIX_HEIGHT")
-        .ok()
-        .and_then(|v| v.parse().ok())
-        .unwrap_or(EVM_GAS_FIX_HEIGHT_DEFAULT)
-}
-
-/// state_root v2 drift fix (2026-05-07, post-halt #5 RCA): pre-fork
-/// `update_trie_for_block` derives `touched_addrs` from `tx.from` /
-/// `tx.to` / `block.validator` + `PROTOCOL_TREASURY` only. That misses
-/// every address mutated by the EVM apply path that isn't named in the
-/// outer Sentrix Tx — CREATE'd contract addresses (computed from
-/// sender + nonce, not in tx.to), internal-CALL recipients (a contract
-/// calling another contract), contract storage from internal SSTOREs.
-/// AccountDB tracks them, the trie doesn't. Validators agree on trie
-/// root for a long time (untracked addresses drift quietly across
-/// hosts) then halt the moment a block touches a previously-drifted
-/// address. Halt at h=1,650,000 on 2026-05-07 localised this exactly:
-/// `state-diff` showed identical block content with divergent trie
-/// roots; STATE-FP traces showed agreeing AccountDB through h=1,649,999.
-///
-/// Post-fork the `touched_addrs` list is augmented with every address
-/// in `accounts.touched_in_block`, the per-block accumulator hooked
-/// into every AccountDB mutator. Default `u64::MAX` keeps the legacy
-/// behaviour unchanged so v2.1.81 chain history stays bit-identical.
-/// Operator flips after testnet bake + chain.db rsync alignment to
-/// pick canonical trie root pre-flip.
-const EXTENDED_TOUCH_LIST_HEIGHT_DEFAULT: u64 = u64::MAX;
-pub fn get_extended_touch_list_height() -> u64 {
-    std::env::var("EXTENDED_TOUCH_LIST_HEIGHT")
-        .ok()
-        .and_then(|v| v.parse().ok())
-        .unwrap_or(EXTENDED_TOUCH_LIST_HEIGHT_DEFAULT)
-}
-
-/// Strict-justification fork (2026-05-07, post-halt #9 RCA): pre-fork
-/// the peer-justification gate verified only stake-weight ARITHMETIC —
-/// it summed the `stake_weight` field on each `SignedPrecommit` (sender-
-/// supplied number) and checked that against the receiver's local
-/// `total_active_stake` threshold. The signatures themselves were never
-/// recovered or matched against the claimed validator. A peer with a
-/// drifted active-set view could send a block whose precommits were
-/// either unsigned, signed by the wrong key, or weighted from a
-/// different registry snapshot, and the receiver would accept the block
-/// silently — applied a fork. Plus the `total_active_stake==0` branch
-/// at simul-start COMPLETELY bypassed the gate, accepting any
-/// justification during the cold-start window.
-///
-/// Post-fork: every precommit signature is recovered with
-/// `Precommit::signing_payload_for_height` + `recover_signer` and
-/// matched against the claimed validator address. Verified-stake is
-/// summed using the receiver's OWN `stake_registry.get_validator(...)`,
-/// not the sender's `stake_weight` field. Threshold gate uses the
-/// receiver's active set; if `total_active_stake==0` post-fork the
-/// gate REJECTS instead of bypassing (cold-start nodes catch up via
-/// `bypass_authz` replay anyway). Default `u64::MAX` preserves
-/// legacy behaviour bit-identically.
-const STRICT_JUSTIFICATION_HEIGHT_DEFAULT: u64 = u64::MAX;
-pub fn get_strict_justification_height() -> u64 {
-    std::env::var("STRICT_JUSTIFICATION_HEIGHT")
-        .ok()
-        .and_then(|v| v.parse().ok())
-        .unwrap_or(STRICT_JUSTIFICATION_HEIGHT_DEFAULT)
-}
+// Fork-height accessors live in `crate::fork_heights` now — re-export so
+// every existing caller path (`crate::blockchain::get_*_height(...)`) keeps
+// resolving bit-identically.
+pub use crate::fork_heights::{
+    get_add_self_stake_height, get_bft_gate_relax_height, get_evm_fork_height,
+    get_evm_gas_fix_height, get_evm_value_transfer_height, get_extended_touch_list_height,
+    get_jail_consensus_height, get_nft_tokenop_height, get_reward_v2_fork_height,
+    get_strict_justification_height, get_tokenomics_v2_height, get_voyager_fork_height,
+    warn_if_jail_consensus_armed,
+};
 
 /// Read chain_id from SENTRIX_CHAIN_ID env var, fallback to 7119.
 pub fn get_chain_id() -> u64 {
